requirements.txt
python-telegram-bot==20.5
SQLAlchemy==1.4.51
apscheduler==3.9.1
python-dotenv==1.0.0
pandas==2.0.3
python-dateutil==2.8.2
httpx==0.24.0
jdatetime